requirements.txt
alembic==1.16.4
bcrypt==4.3.0
blinker==1.9.0
cached-property==2.0.1
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2

mypy_extensions==1.1.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.3
six==1.17.0
SQLAlchemy==2.0.41
sqlalchemy-orm==1.2.10
typing-inspect==0.9.0
typing_extensions==4.14.1
Werkzeug==3.1.3
WTForms==3.2.1
